Adapted PrimaryButton to become MainButton
