+++ front-end/src/index.scss
@@ -1,4 +1,3 @@
-// Fonts
 @font-face {
   font-family: 'Mattone';
   src: local('Mattone'), url(./fonts/Mattone-Regular.otf) format('truetype');
